+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Allow defining default query parameters with the `_query` route default
+ * Add `RequestContext::runWith()` to generate and match URLs for another host, scheme or base URL without leaking the change
 
 8.1
 ---